Update project-planning.md
Tweaks at the top to clarify what this section is about
Tweaks to "transport" and "catering" to try to make it feel more cohesive
But I don't think "Project planning" and "Decision trees" are the best titles for the 10 November launch milestone. Perhaps combine into one section - we can add start a "Decision Trees" section when we have our first decision tree? Or keep two sections, but with a clearer conceptual division, e.g. "Applying for funding" (this one) and "Running a research project"? Not sure!